+++ src/applications/caregivers/config/chapters/secondaryOne/hasSecondaryOneCaregiver.js
@@ -1,21 +1,43 @@
 import definitions from 'applications/caregivers/definitions/caregiverUI';
-import { secondaryCaregiverFields } from 'applications/caregivers/definitions/constants';
-import { SecondaryRequiredAlert } from 'applications/caregivers/components/AdditionalInfo';
-import { shouldHideAlert } from 'applications/caregivers/helpers';
+import {
+  primaryCaregiverFields,
+  secondaryCaregiverFields,
+} from 'applications/caregivers/definitions/constants';
+import { SecondaryCaregiverInfo } from 'applications/caregivers/components/AdditionalInfo';
 
-const { hasSecondaryCaregiverOneUI } = definitions.sharedItems;
+const { secondaryRequiredAlert } = definitions.sharedItems;
 
-const hasSecondaryOneCaregiverPage = {
+const hasSecondaryCaregiverPage = {
   uiSchema: {
-    [secondaryCaregiverFields.secondaryOne
-      .hasSecondaryCaregiverOne]: hasSecondaryCaregiverOneUI,
-    'view:secondaryAlert': {
-      'ui:title': ' ',
-      'ui:widget': SecondaryRequiredAlert,
-      'ui:options': {
-        hideIf: formData => shouldHideAlert(formData),
-      },
+    [secondaryCaregiverFields.secondaryOne.hasSecondaryCaregiverOne]: {
+      'ui:title': 'Would you like to add a Secondary Family Caregiver?',
+      'ui:widget': 'yesNo',
+      'ui:description': SecondaryCaregiverInfo({
+        additionalInfo: true,
+        headerInfo: false,
+      }),
+      'ui:required': formData =>
+        !formData[primaryCaregiverFields.hasPrimaryCaregiver],
+      'ui:validations': [
+        {
+          validator: (errors, fieldData, formData) => {
+            const hasPrimary =
+              formData[primaryCaregiverFields.hasPrimaryCaregiver];
+            const hasSecondary =
+              formData[
+                secondaryCaregiverFields.secondaryOne.hasSecondaryCaregiverOne
+              ];
+            const hasCaregiver = hasPrimary || hasSecondary;
+
+            if (!hasCaregiver) {
+              // We are adding a blank error to disable the ability to continue the form but not displaying the error text its self
+              errors.addError(' ');
+            }
+          },
+        },
+      ],
     },
+    'view:secondaryAlert': secondaryRequiredAlert,
   },
   schema: {
     type: 'object',
@@ -30,4 +52,4 @@ const hasSecondaryOneCaregiverPage = {
   },
 };
 
-export default hasSecondaryOneCaregiverPage;
+export default hasSecondaryCaregiverPage;
