+++ src/applications/caregivers/definitions/caregiverUI.js
@@ -10,18 +10,23 @@ import {
   PleaseSelectVAFacility,
   VeteranSSNInfo,
   SecondaryRequiredAlert,
-  PrimaryCaregiverInfo,
   SecondaryCaregiverInfo,
+  AdditionalCaregiverInfo,
 } from 'applications/caregivers/components/AdditionalInfo';
 import { createUSAStateLabels } from 'platform/forms-system/src/js/helpers';
 import { states } from 'platform/forms/address';
 import get from 'platform/utilities/data/get';
-import { vetFields } from './constants';
+import {
+  vetFields,
+  primaryCaregiverFields,
+  secondaryCaregiverFields,
+} from './constants';
 import {
   medicalCenterLabels,
   medicalCentersByState,
   validateSSNIsUnique,
   facilityNameMaxLength,
+  shouldHideAlert,
 } from 'applications/caregivers/helpers';
 
 const emptyFacilityList = [];
@@ -98,40 +103,6 @@ export default {
       'ui:title': ' ',
       'ui:description': AdditionalCaregiverInfo,
       'ui:widget': 'yesNo',
-    },
-    hasPrimaryCaregiverOneUI: {
-      'ui:title': 'Would you like to add a Primary Family Caregiver?',
-      'ui:description': PrimaryCaregiverInfo({
-        additionalInfo: true,
-        headerInfo: false,
-      }),
-      'ui:widget': 'yesNo',
-    },
-    hasSecondaryCaregiverOneUI: {
-      'ui:title': 'Would you like to add a Secondary Family Caregiver?',
-      'ui:description': SecondaryCaregiverInfo({
-        additionalInfo: true,
-        headerInfo: false,
-      }),
-      'ui:widget': 'yesNo',
-      'ui:required': formData =>
-        !formData[primaryCaregiverFields.hasPrimaryCaregiver],
-      'ui:validations': [
-        {
-          validator: (errors, fieldData, formData) => {
-            const hasPrimary =
-              formData[primaryCaregiverFields.hasPrimaryCaregiver];
-            const hasSecondary =
-              formData[primaryCaregiverFields.hasSecondaryCaregiverOne];
-            const hasCaregiver = hasPrimary || hasSecondary;
-
-            if (!hasCaregiver) {
-              // We are adding a blank error to disable the ability to continue the form but not displaying the error text its self
-              errors.addError(' ');
-            }
-          },
-        },
-      ],
     },
   },
   vetUI: {
